+++ cmd/deployment-splitter/main.go
@@ -22,7 +22,7 @@ import (
 	"k8s.io/client-go/tools/clientcmd"
 	"k8s.io/klog/v2"
 
-	"github.com/kcp-dev/kcp/pkg/reconciler/deployment"
+	"github.com/kcp-dev/kcp/pkg/reconciler/workload/deploymentsplitter"
 )
 
 const numThreads = 2
@@ -45,5 +45,5 @@ func main() {
 		klog.Fatal(err)
 	}
 
-	deployment.NewController(r).Start(numThreads)
+	deploymentsplitter.NewController(r).Start(numThreads)
 }
