+++ pkg/reconciler/workload/basecontroller/controller.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package cluster
+package basecontroller
 
 import (
 	"context"
