+++ pkg/reconciler/workload/deploymentsplitter/deploymentsplitter_controller.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package deployment
+package deploymentsplitter
 
 import (
 	"context"
