+++ pkg/reconciler/workload/heartbeat/heartbeat_start.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package workloadclusterheartbeat
+package heartbeat
 
 import (
 	"fmt"
