+++ pkg/reconciler/workload/namespace/conditions.go
