+++ pkg/reconciler/workload/syncer/pullmanager.go
