+++ pkg/reconciler/workload/syncer/syncer_controller_start.go
