+++ pkg/syncer/specsyncer.go
@@ -33,7 +33,7 @@ import (
 	"k8s.io/klog/v2"
 	"k8s.io/utils/pointer"
 
-	nscontroller "github.com/kcp-dev/kcp/pkg/reconciler/namespace"
+	nscontroller "github.com/kcp-dev/kcp/pkg/reconciler/workload/namespace"
 )
 
 func deepEqualApartFromStatus(oldObj, newObj interface{}) bool {
